Add files via upload
Using machine learning in production process to ensure that every wine bottle has the highest quality. This will in turn reduce the cost with regard to each wine bottle as less human labor is required for monitoring of quality. Therefore, this analysis is done to predict wine quality given some important attributes in wine.